added report generating

+++ spam_detector/fitting.py
@@ -1,6 +1,8 @@
 import os
+import sys
 
 import warnings
+from typing import NamedTuple
 
 import nltk
 import numpy as np
@@ -17,18 +19,48 @@ from sklearn.naive_bayes import MultinomialNB
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.svm import SVC
 
-from sklearn.metrics import  accuracy_score, confusion_matrix
+from sklearn.metrics import accuracy_score, confusion_matrix
 from nltk.corpus import stopwords
 
 from gensim.models import Word2Vec
 from gensim.models import FastText
 from nltk.tokenize import word_tokenize
 
+from dataclasses import dataclass
+
 from .utils import vectorize
 
 import shutil
 
 warnings.simplefilter(action='ignore', category=FutureWarning)
+
+
+@dataclass
+class VectorizerFitReport:
+    id: str | None
+    lang: str
+    title: str
+    duration: float
+    size: int
+
+
+@dataclass
+class ClassifierModelMetrics:
+    TN: int
+    FP: int
+    FN: int
+    TP: int
+
+
+@dataclass
+class ClassifierFitReport:
+    lang: str
+    title: str
+    vectorizer_id: str
+    duration: float
+    size: int
+    metrics: ClassifierModelMetrics
+    accuracy: float
 
 
 def process(text):
@@ -38,7 +70,7 @@ def process(text):
     return clean
 
 
-def fit_MNB(dataset):
+def fit_bag_of_words(dataset):
     model_bag_of_words = CountVectorizer(analyzer=process)
     message_bag_of_words = model_bag_of_words.fit_transform(dataset['text'])
     # %%
@@ -57,13 +89,15 @@ def fit_fast_text(tokenized_text):
 
 
 def create_model_dir_struct():
-    assert os.path.exists("models") and os.path.isdir("models"),"models directory doesn't exist!"
-    for filename in os.listdir("models"):
-        file_path = os.path.join("models", filename)
-        if os.path.isfile(file_path):
-            os.remove(file_path)
-        elif os.path.isdir(file_path):
-            shutil.rmtree(file_path)
+    if os.path.exists("models") and os.path.isdir("models"):  # , "models directory doesn't exist!"
+        for filename in os.listdir("models"):
+            file_path = os.path.join("models", filename)
+            if os.path.isfile(file_path):
+                os.remove(file_path)
+            elif os.path.isdir(file_path):
+                shutil.rmtree(file_path)
+    else:
+        os.mkdir("models")
     os.chdir("models")
     os.makedirs("ru")
     os.makedirs("en")
@@ -79,35 +113,106 @@ def create_model_dir_struct():
     os.chdir("..")
 
 
+def form_md_report(reports: list[VectorizerFitReport | ClassifierFitReport]) -> str:
+    def vectorizer_report_item(vfr: VectorizerFitReport) -> str:
+        return f"- {vfr.title}, duration of fitting: {vfr.duration} seconds\n"
+
+    def classifier_report_item(cfr: ClassifierFitReport) -> str:
+        table = PrettyTable()
+        table.field_names = ["Predicted \\ Actual", "Positive", "Negative"]
+        table.add_row(["Positive", cfr.metrics.TP, cfr.metrics.FN])
+        table.add_row(["Negative", cfr.metrics.FP, cfr.metrics.TN])
+
+        return f"""
+#### {cfr.title} by {cfr.vectorizer_id}
+
+Duration of fitting: {cfr.duration} seconds
+
+```                       
+{table}
+```
+                       
+Accuracy of {cfr.title}-> {cfr.accuracy}%
+"""
+
+    return f"""
+# Report
+    
+## English
+    
+### Vectorizers
+    
+{''.join(map(vectorizer_report_item,
+             filter(lambda report: isinstance(report, VectorizerFitReport) and report.lang == "en",
+                    reports)))}
+    
+### Classifiers
+    
+{''.join(map(classifier_report_item,
+             filter(lambda report: isinstance(report, ClassifierFitReport) and report.lang == "en",
+                    reports)))}
+    
+## Russian
+    
+### Vectorizers
+    
+{''.join(map(vectorizer_report_item,
+             filter(lambda report: isinstance(report, VectorizerFitReport) and report.lang == "ru",
+                    reports)))}
+    
+### Classifiers
+    
+{''.join(map(classifier_report_item,
+             filter(lambda report: isinstance(report, ClassifierFitReport) and report.lang == "ru",
+                    reports)))}
+    
+"""
+
+
 def prepare_and_fit(datasets=None):
     if datasets is None:
         datasets = {"en": 'dataset/emails_en.csv', "ru": 'dataset/emails_ru.csv'}
-    nltk.download('stopwords')
-    nltk.download('punkt')
     create_model_dir_struct()
-    for lang in ["ru", "en"]:  # "ru"
-        print(f"fitting the model by lang={lang}")
+    reports = list()
+    for lang in ["ru", "en"]:
         dataset = pd.read_csv(datasets[lang])
+        # dataset = pd.concat([dataset[dataset["spam"] == 1].head(100), dataset[dataset["spam"] == 0].head(100)])
         # Check for duplicates and remove them
         dataset.drop_duplicates(inplace=True)
         # Fit the CountVectorizer to data
         tokenized_text = [word_tokenize(text) for text in dataset['text']]
-
         start_time = time.time()
-        model_bag_of_words, message_bag_of_words = fit_MNB(dataset)
+        model_bag_of_words, message_bag_of_words = fit_bag_of_words(dataset)
         finish_time = time.time()
-        print(f"\tfitting bag of words, duration: {finish_time - start_time} seconds")
+        reports.append(VectorizerFitReport(
+            id="bag_of_words",
+            lang=lang,
+            title="Bag of words",
+            duration=finish_time - start_time,
+            size=sys.getsizeof(model_bag_of_words)
+        ))
 
         start_time = time.time()
         model_word2text, message_word2vec = fit_word2text(tokenized_text)
         finish_time = time.time()
-        print(f"\tfitting word to vec, duration: {finish_time - start_time} seconds")
+        reports.append(VectorizerFitReport(
+            id="word2vec",
+            lang=lang,
+            title="Word to vector",
+            duration=finish_time - start_time,
+            size=sys.getsizeof(model_word2text),
+        ))
 
         start_time = time.time()
         model_fast_text, message_fast_text = fit_fast_text(tokenized_text)
         finish_time = time.time()
-        print(f"\tfitting fast text, duration: {finish_time - start_time} seconds")
-
+        reports.append(VectorizerFitReport(
+            id="fast_text",
+            lang=lang,
+            title="Fast Text",
+            duration=finish_time - start_time,
+            size=sys.getsizeof(model_fast_text),
+        ))
         minimal = min([min(vec) for vec in message_fast_text])
         if minimal < 0:
             minimal = np.abs(minimal)
@@ -124,8 +229,6 @@ def prepare_and_fit(datasets=None):
         dump(model_bag_of_words, open(f"models/{lang}/vectorizers/bag_of_words_vectorizer.pkl", "wb"))
         dump(model_fast_text, open(f"models/{lang}/vectorizers/fast_text_vectorizer.pkl", "wb"))
         dump(model_word2text, open(f"models/{lang}/vectorizers/word2vec_vectorizer.pkl", "wb"))
-        finish_time = time.time()
-        print(finish_time - start_time)
         for vectorized_message, vectorizer_name in [
             (message_bag_of_words, "bag_of_words"),
             (message_fast_text_non_negative, "fast_text"),
@@ -133,44 +236,43 @@ def prepare_and_fit(datasets=None):
         ]:
             x_train, x_test, y_train, y_test = train_test_split(vectorized_message, dataset['spam'], test_size=0.20,
                                                                 random_state=0)
-            print(vectorizer_name)
-            for Model, model_name in [
-                (MultinomialNB, "MultinomialNB"),
-                (RandomForestClassifier, "RandomForestClassifier"),
-                (SVC, "SVC"),
+            for Model, model_id, model_name in [
+                (MultinomialNB, "MultinomialNB", "Naive Bayes"),
+                (RandomForestClassifier, "RandomForestClassifier", "Random forest classifier"),
+                (SVC, "SVC", "Support Vector Classification"),
             ]:
-                print(f"\n\n\n<= {lang}:{vectorizer_name}:{model_name} =>\n")
                 try:
-
                     model = Model()
 
                     start_time = time.time()
                     model.fit(x_train, y_train)
                     finish_time = time.time()
-                    print(f"fitting {model_name} by {vectorizer_name}: {finish_time - start_time} seconds")
                     # Model predictions on test set
                     y_pred = model.predict(x_test)
                     cm = confusion_matrix(y_test, y_pred)
-                    table = PrettyTable()
-                    print(f"TP = {cm[0][0]}\nFP = {cm[1][0]}\nTN = {cm[0][1]}\nFN = {cm[1][1]}\n")
-
-                    table.field_names = ["Predicted \\ Actual", "Positive", "Negative"]
-                    # добавление данных по одной строке за раз
-                    table.add_row(["Positive", cm[0][0], cm[1][0]])
-                    table.add_row(["Negative", cm[0][1], cm[1][1]])
-
-                    print(f"{table}")
+                    tn, fp, fn, tp = cm.ravel()
 
                     # Model Evaluation | Accuracy
                     accuracy = accuracy_score(y_test, y_pred)
 
-                    print(f"accuracy of {model_name}-> {accuracy * 100}%")
                     # Model Evaluation | Classification report
 
-                    dump(model, open(f"models/{lang}/classifiers/{vectorizer_name}/{model_name}.pkl", 'wb'))
-
-                    # print(classification_report(y_test, y_pred))
+                    reports.append(ClassifierFitReport(
+                        lang=lang,
+                        title=model_name,
+                        vectorizer_id=vectorizer_name,
+                        duration=finish_time - start_time,
+                        size=sys.getsizeof(model),
+                        metrics=ClassifierModelMetrics(
+                            TP=tp, TN=tn, FP=fp, FN=fn,
+                        ),
+                        accuracy=accuracy * 100,
+                    ))
+                    dump(model, open(f"models/{lang}/classifiers/{vectorizer_name}/{model_id}.pkl", 'wb'))
 
                 except Exception as e:
-                    print(f"\t{model_name} ->"
+                    print(f"\t{model_name} of {vectorizer_name} ->"
                           f"\n\t {e}")
+
+    with open("./report.md", "w") as report_file:
+        report_file.write(form_md_report(reports))
